ECS - Update LifetimeComponent and implement LifetimeSystem
Changed `timeLeft` from long (milliseconds) to float (seconds) and implemented `LifetimeSystem` to decrement entity lifetime and remove expired entities from the engine.

+++ core/src/main/java/com/fatpiggies/game/model/ecs/components/LifetimeComponent.java
@@ -3,5 +3,5 @@ package com.fatpiggies.game.model.ecs.components;
 import com.badlogic.ashley.core.Component;
 
 public class LifetimeComponent implements Component {
-    public long timeLeft; // Milliseconds
+    public float timeLeft; // Seconds for deltaTime
 }
